fix prioritization scores

+++ experimental/quick-check-2023/src/lib/utils.js
@@ -23,5 +23,5 @@ export const recode_numeric_range = (
 };
 
 export const arrayAverage = (array) => {
-    if (array.length) { return array.reduce((a, b) => parseInt(a) + parseInt(b)) / array.length } else { return null }
+    if (array.length) { return array.reduce((a, b) => parseFloat(a) + parseFloat(b)) / array.length } else { return null }
 }
